test(ci): delete the YAML-regex change-detector tests, keep the script-execution tests (#66)
Drop the 8 tests that asserted against workflow YAML and package.json
content -- they failed whenever the YAML was edited, not when it was
wrong, and the CI overhaul is all edits. The 3 script-execution tests
survive in scripts/__tests__/consumer-scripts.test.mjs, renamed to
match its remaining content.

Each invariant that survives the gate map is preserved as a comment
beside the line it guards: per-event concurrency groups (ci.yml,
docs.yml; exact-package-candidate.yml already had one), detox
build-framework-cache ordering, Gradle invoked from the generated
consumer directory, and the Jest self-reference mapping in
package.json.
