+++ .github/workflows/docs.yml
@@ -13,6 +13,8 @@ permissions:
   contents: read
 
 concurrency:
+  # Keep the event name in the group: sharing one group across events once let
+  # a release-authorizing push run deadlock behind other events for 13 hours.
   group: docs-${{ github.event_name }}-${{ github.ref }}
   cancel-in-progress: true
